Scheduler2.0: 스케줄러 경험치 마이너스 테스트

+++ src/main/java/com/sollertia/habit/domain/completedhabbit/dto/SimpleHabitVo.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 public class SimpleHabitVo {
     private String title;
     private Long achievement;
-    private int accomplishCount;
+    private Long accomplishCount;
     private Long goalCount;
 
     private String startDate;
